move CartIconCounterComponent to the separate folder
