+++ src/components/cart-icon-counter/CartIconCounter.component.tsx
